requirements/requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.9
djangorestframework==3.14.0
drf-yasg==1.21.7
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-pubsub==2.21.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
inflection==0.5.1
mock==5.1.0
mysqlclient==2.2.3
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
structlog==24.1.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
